Allow sessreg to be built without <utmp.h>.
As of 9.0, FreeBSD will only support <utmpx.h>.  It only implements the
POSIX interface with minor extensions.  There is no need to write to
wtmp and lastlog separately, because this is already done by
pututxline().  Add additional checks to configure.ac to search for the
presence of utmpxname() and updwtmpx().

Signed-off-by: Ed Schouten <ed@80386.nl>
Signed-off-by: Alan Coopersmith <alan.coopersmith@sun.com>