Remove dubious code.
The NO_UTMP definition is set on non-FreeBSD, non-OpenBSD, BSD-based
systems. When looking at the commonly used BSD-based operating systems,
they either use utmpx, or I can't think of a reason why the utmp code
wouldn't work.

If it turns out some obscure operating system breaks because of this
change, we should replace this by something more accurate, such as an
Autoconf check or an #ifdef specific to that operating system.

Signed-off-by: Ed Schouten <ed@80386.nl>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>